+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add allowNullValues boolean option to pass Null values through helm configs without having them scrubbed (https://github.com/pulumi/pulumi-kubernetes/issues/2089)
+- Fix replacement behavior for immutable fields in SSA mode (https://github.com/pulumi/pulumi-kubernetes/issues/2232)
 
 ## 3.22.1 (October 26, 2022)